requirements.txt
alabaster==0.7.13
babel==2.16.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
docutils==0.20.1
exceptiongroup==1.2.2
future==1.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.5
junos-eznc==2.7.2
lxml==5.3.0
MarkupSafe==2.1.5
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.4.0
netutils==1.12.0
ntc_templates==7.6.0
packaging==24.2
paramiko==3.5.0
pluggy==1.5.0
pycparser==2.22
pyeapi==1.0.4
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.1.4
pyserial==3.5
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
scp==0.15.0
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
textfsm==1.1.3
tomli==2.2.1
transitions==0.9.2
ttp==0.9.5
ttp-templates==0.3.7
typing_extensions==4.12.2
urllib3==2.2.3
yamlordereddictloader==0.4.2
zipp==3.20.2
